core: RPMB_FS: remember owner uuid
Prior to this patch was the owning uuid looked up via the current
session.  The assumption that the uuid of the context of the current
session is the same as the owner of a RPMB file doesn't hold any longer
after:
commit 078f18f82eeb ("core: RPMB FS: provide tee_rpmb_fs_raw_open()")
was introduced.

Instead the pointer to the uuid passed as the owner is stored in the
file handle and used as needed when encrypting/decrypting the file.

Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
Fixes: https://github.com/OP-TEE/optee_os/issues/1780
Tested-by: Jens Wiklander <jens.wiklander@linaro.org> (QEMU)
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>