core: allocate ns buffer for stmm
Current StandaloneMM implementation works on a static non-secure buffer.
StMM saves this buffer region on initialization. An MM Communicate call
must be inside this region. This commit allocates an NS buffer for StMM.

Since OP-TEE can't ensure parameter memmapping to land inside the
region. After it is mapped, it is then memcpy'ed to the buffer.

Signed-off-by: Pipat Methavanitpong <pipat.methavanitpong@linaro.org>