console: lx2160A: fix uart clock and baud rate
Currently there are garbled messages from OP-TEE due to assumption that
platform clock is always 700MHz. This is not correct as LX2 supports
variable platform frequency. It could be one of the 600, 650, 700, 750 Mhz
based on the RCW configuration.

Ideally OPTEE should read RCW registers from Global Utilities Register block
and derive the uart clock based on platform pll frequency. But there is no need
for this as Baud Rate is already configured in PL011 by the previous boot stages
in TF-A.  This fix calls pl011_init() with zero for baud rate => It won't be
reinitalized in OP-TEE.

Signed-off-by: Thirupathaiah Annapureddy <thiruan@linux.microsoft.com>
Signed-off-by: Sahil Malhotra <sahil.malhotra@nxp.com>